+++ CommunityCenterOverhaul/Converter.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace CommunityCenterBundleOverhaul
+{
+    public class Converter
+    {
+        public static readonly JsonSerializerSettings Settings = new JsonSerializerSettings
+        {
+            MetadataPropertyHandling = MetadataPropertyHandling.Ignore,
+            DateParseHandling = DateParseHandling.None,
+        };
+    }
+}
